did css styling to log in page and collections page. Renamed table to AllCards.
