+++ src/Components/Header/Header.css
@@ -1,0 +1,32 @@
+.pageTitle {
+    text-align: center;
+}
+
+.userInfo {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    position: absolute;
+    right: 40px;
+    top: 30px;
+}
+
+.username {
+    padding: 10px;
+}
+
+.profilePic {
+    border: 1px solid black;
+    border-radius: 50%;
+    height: 100px;
+    width: 100px;
+}
+
+.logInOut {
+    width: 70px;
+}
+
+.logInLink {
+    color: inherit;
+    text-decoration: none;
+}
