+++ src/Components/Header/Header.js
@@ -1,12 +1,14 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
 import NavBar from "../NavBar/NavBar";
+import { Link } from "react-router-dom";
+import "./Header.css";
 
 export default function Header(props) {
     const { setLoggedInUser, loggedInUser, loggedOut } = props;
     return (
         <>
-            <h1>Pokemon Card Collector</h1>
+            <h1 className="pageTitle">Pokemon Card Collector</h1>
             <NavBar />
 
             <div className="userInfo">
@@ -15,15 +17,25 @@ export default function Header(props) {
                     src={loggedInUser.profilePic}
                     alt="profile pic for user"
                 />
-                <span className="username">{loggedInUser.username}</span>
-                {loggedInUser.userID > 0 && (
+                {loggedInUser.username ? (
+                    <span className="username">{loggedInUser.username}</span>
+                ) : (
+                    <span className="username">Guest User</span>
+                )}
+                {loggedInUser.userID > 0 ? (
                     <button
-                        className="logOut"
+                        className="logInOut"
                         onClick={() => {
                             setLoggedInUser(loggedOut);
                         }}
                     >
                         Log Out
+                    </button>
+                ) : (
+                    <button className="logInOut">
+                        <Link className="logInLink" to="/logIn">
+                            Log In
+                        </Link>
                     </button>
                 )}
             </div>
