+++ src/Pages/LogInPage/LogInPage.css
@@ -1,0 +1,5 @@
+.LogInCreateUser {
+    padding-top: 100px;
+    margin: auto;
+    width: 650px;
+}
