+++ src/Pages/LogInPage/LogInPage.js
@@ -1,11 +1,12 @@
 import React from "react";
-import LogIn from "../Components/LogIn/LogIn";
-import CreateUser from "../Components/CreateUser/CreateUser";
+import LogIn from "../../Components/LogIn/LogIn";
+import CreateUser from "../../Components/CreateUser/CreateUser";
+import "./LogInPage.css";
 
 export default function LogInPage(props) {
     const { loggedInUser, updateFormData, users, setLoggedInUser } = props;
     return (
-        <>
+        <div className="LogInCreateUser">
             {loggedInUser.userID < 1 && (
                 <LogIn
                     updateLogInRequest={updateFormData}
@@ -14,6 +15,6 @@ export default function LogInPage(props) {
                 />
             )}
             <CreateUser updateUserData={updateFormData} users={users} />
-        </>
+        </div>
     );
 }
